First refactoring
