+++ src/components/Divider.jsx
@@ -1,7 +1,4 @@
 
-function Divider() {
+export function Divider() {
     return <div style={{ height: '4px', width: '150%', borderBottom: '4px dashed #E5E7EB' }} />;
 }
-
-
-export default Divider;
