+++ src/pages/ErrorPage.jsx
@@ -1,8 +1,9 @@
 
+// This components is displayed if the user enters a non-existing url
 export default function ErrorPage() {
 
   return (
-    <div id="error-page">
+    <div id="error-page" className="flex itemes-center justify-center h-screen">
       <h1 className="">Oops!</h1>
       <p>Lo sentimos, un error inesperado a ocurrido.</p>
       <p>
